Refactors X and Y to Data & Column
Removes the generic X and Y reference, and renames the axis to data and column axis.
All the instances of axis are renamed to dataAxis and columnAxis, but the variables x and y are left as it is, as they are heavily used in calculations in the Line and Column files.

Signed-off-by: Anmol Jagetia <anmoljagetia@gmail.com>